Refactor: Use one namespace Update: xUnit Use of ICollectionFixture
