+++ Validators.Tests/Tests/Validators/Postcode/OnlyDigitsTests.cs
@@ -4,10 +4,10 @@ using Validators.Interfaces;
 using Xunit;
 
 
-namespace Validators.Tests.Postcode.Tests
+namespace Validators.Tests.Postcode
 {
+    [Collection("Postcodes")]
     public class OnlyDigitsTests
-        : IClassFixture<PostcodeFixture>
     {
 
         private readonly IPostcodeValidator _postcodeValidator;
